ExifTool Support:
Add new option to write metadata to files using ExifTool nackend instead Exiv2.
This option replace all calls to Exiv2 to patch files by ExifTool as well.
This will prevent to corrupt files due to lack of stability support from Exiv2.
This includes large XMP section in JPEG and MPO files.
For RAW and DNG files 2 options still here but are depends of ExifTool availability.
We will never write at all to DNG or RAW files with Exiv2.
This new option to replace Exiv2 by ExifTool is turned off by default for the moment,
but if no side-effects is discovered in the future before official 8.0.0 release,
this option will be turned on by default to be safe in all cases.

BUGS: 325458
BUGS: 448729

FIXED-IN: 8.0.0