Gsoc22 aesthetic detector
##Description
This is the implementation of aesthetic detector. The class can load model from an pb file which is a pre-trained model of aesthetic classification with 3 classes and calculate the result of the mode 

##Unit test
Unit test on 3 good images from AVA dataset, and 3 bad images that are already in the data of unit test of IQS

##TODO
The management of model file is not implemented yet, the reason is that pb file of the pretrained model is a large file ( 200MB) to store on gitlab repo of digikam. There should be another solution to store or download this file. Temporary, the model can be downloaded from [here](https://drive.google.com/file/d/1c4uVuyLp_eqE1vLCEVgU0LXv1ik5YgA1/view?usp=sharing)