Fix GPUMemCopy's use of SSE.
The availability of SSE2 and SSE4 are detected at compile time.
When built on one machine but run on another, it may happen that
the build machine supports SSE4 but the target machine does not.

In GPUMemCopy, the function isAvailable() supports this by first
checking at run-time for sse4, then sse2, then finally returning false.

This change fixes copyFrame() and gpu_memcpy() to similarly fall back
to SSE2 if available.