timeadjust: handle file time equally on sidecar
Updating metadata directly in an item honors the updateFileTimeStamp
option. If this option is not enabled, file modification time of an item
is preserved (restored) during metadata change.

Updating metadata in a sidecar does NOT honor the updateFileTimeStamp
option. All actions in Digikam seems to always update the file
modification time of a sidecar when it is written, disregarding the
updateFileTimeStamp option. Except the TimeAdjustTask which tries to
keep sidecar file time aligned with the item file time.

One consequences is, if an external tool detects changes to sidecar
files based on file modification time, it correctly reacts to all
changes done in Digikam except for the time adjustment.